scalefs: Avoid double-addition to pending-stack when processing renames
If we find the rename-link or rename-unlink operation, we should add the
counterpart to the pending stack. But when we actually find the counterpart
next, we shouldn't re-add the original operation again. Fix this by adding
appropriate checks.

Also, split up the assert statement in process_metadata_log() into multiple
lines so that it is easier to figure out which specific stack is non-empty
if the system crashes at that point.