scalefs: Bypass processing journal and inodereclaim files on fresh filesystems
For most of our experiments, we flash the filesystem to the disk/memory during
boot, and so there is no reason why we should waste time going through all the
per-cpu journal and inodereclaim files trying to do crash-recovery. Skip those
parts to speedup booting when crash-recovery is known to be unnecessary.