scalefs: Fix sync_dirty_files_and_dirs() to flush a consistent set of inodes
sv6 seems to be crashing at sync_dirty_files_and_dirs() when running some of the
Commuter test-cases because it is not able to find the inode mapping for some of
the mnodes that it enumerates from the metadata-log hash-table. But just before
we call sync_dirty_files_and_dirs(), we invoke process_metadata_log() which also
enumerates mnodes from the same hash-table.

So this seems to be a race where new mnodes get added to the hash-table in
between these 2 calls, and sync_dirty_files_and_dirs() has trouble dealing with
them, since the inodes for those mnodes may not have been allocated yet. Fix
this by passing the mnode-number list that we built from process_metadata_log()
to sync_dirty_files_and_dirs(), as those mnodes are guaranteed to have their
inodes in place. Besides, sync makes no promise that there won't be any dirty
files or directories when it returns, if the user goes on dirtying files and
directories in parallel to the sync() call.

Furthermore, we can also perform just one flush_journal() call for both of these
functions, instead of doing it twice, by making use of the ordering property
that we guarantee within a given per-cpu journal transaction queue.