scalefs: Mark mnode as clean before deleting its inode
If a file is created and unlinked from a directory, and the directory is
fsynced, the absorption code might kick in and delete the inode corresponding to
the file's mnode. After that, there is no point in fsyncing that file mnode's
data-blocks even if they are dirty.

In the current code, we only mark the file's mnode as clean when we fsync its
data-blocks. But as we can see in this scenario, we should also mark it as clean
before deleting it via absorption; otherwise, sync_file_page() will crash because
it won't be able to find the inode corresponding to the dirty mnode it is
trying to flush.