scalefs, diskfs: Add dirty blocks to transactions lazily
The LFS Largefile benchmark creates a file and writes 100 MB of data to it
sequentially (and later on, at random offsets). When performing an fsync after
this file growth, the direct/indirect/doubly-indirect blocks are updated for
every data block that is allocated and written to the disk.

The current code copies the dirty metadata blocks immediately to the
transaction, which means that for a 100 MB file growth, the fsync incurs an
unnecessary 100 MB block-copy overhead (in-memory) by copying the modified
metadata blocks over and over; but in reality, it turns out that only 28 of
those metadata blocks are unique!

Optimize this by just noting down which metadata blocks are dirty, and then,
after all the writes are completed, copy the unique metadata blocks into the
transaction lazily. [However, we must take care to hold the ilock over the file
throughout; otherwise we risk logging incorrect snapshots of the blocks into
the transaction].