scalefs: Distinguish auxiliary transactions used in the journal commit code
We use a separate transaction called prune_trans to deduplicate blocks across
multiple individual transactions. This is also used to apply the blocks of all
the individual transactions to the on-disk filesystem after commit. We use a
different transaction to write the transaction blocks along with the appropriate
journal headers to the on-disk journal. Rename the latter to 'jrnl_trans' to
indicate its purpose a bit more clearly.