scalefs: Acquire inode-block lock when fsyncing files
The current code fsyncs files in 2 parts: first it flushes the create operation,
thus creating an inode for that file on the disk. This acquires the appropriate
locks correctly. Next, the file's data blocks are updated (without using
transactions), and the updates to the file's metadata blocks are flushed to disk
in a new transaction. In this latter part, we had forgot to acquire the
appropriate inode-block lock. Fix it.