ahci: Make all the 32 command slots work
At the moment, only 4 of the 32 command slots work in the AHCI driver
(specifically, slots whose index is divisible by 8). Using any other command
slot results in a hang.

The current code allocates 17 prdt entries per command slot (65536/4096 + 1).
Allocating 8 prdt entries instead, or 65536 entries (as we do in this patch)
happens to fix the above problem! That is, with this change, we can use all
the 32 command slots!

However, the results are not all that encouraging. Using only 1 command slot
drops the throughput of LFS largefile benchmark by nearly 30% (as compared to
using 4 command slots). But using all the 32 command slots doesn't improve the
throughput any further beyond that with 4 command slots.