scalefs: Don't apply transactions in the fsync path
An fsync() doesn't have to apply the transaction to the on-disk filesystem. It
is sufficient if it merely commits the transaction to the on-disk journal.

So optimize fsync() by not applying the transactions to the disk in the normal
case, and applying them only when the journal is full. [Sync will continue to
apply on every invocation, so that it returns with no work left pending].

However, make arrangements to satisfy cross-queue (cross-journal) dependencies
between transactions waiting to be applied, by force-applying the dependent
transactions if necessary.

This reduces the number of calls to ideflush() and also enables better batching
and deduplication among transactions.