scalefs: Flush 'create' operation along with the fsync transaction whenever possible
When fsyncing a freshly created file (or directory), we can actually flush the
'create' operation as well as the fsync transaction (which modifies the file's
inode contents) in a single transaction batch, instead of committing them
individually. Add this optimization to help further reduce the number of calls
to ideflush().