scalefs: Implement an efficient and leaner layout for the on-disk journal
The current layout for committing transactions to the on-disk journal is not
very efficient - it consumes more space (blocks) per-transaction than necessary
and more importantly, committing a transaction to the journal also involves
read-modify-writes of blocks in the journal, because we commit the transaction
using several small writes which are neither aligned to block boundaries nor
guaranteed to be of block-size length. As expected, this slows down transaction
commit due to the disk reads involved in the read-modify-writes of the journal.

Ideally, we should never have to read anything back from the journal except
during crash-recovery. So this patch introduces a completely re-written, efficient
layout for the on-disk journal, which gets rid of all read-modify-writes and
also fits every transaction commit into (almost) the minimum number of blocks it
needs.

With this layout, every transaction committed in the journal looks like this:

Start Block | (Optional) Address Block | Data Blocks | Commit Block

The start block, the address block and the commit blocks are all exactly block-size
in length, and there are 'n' data blocks which are also each of length block-size.
Thus, every transaction begins and ends at a block boundary, and hence the
journal is never read because we always overwrite entire blocks with new info.

The data blocks are the blocks from the transaction itself, and hence we need
some metadata in the commit to guide the crash-recovery routine as to where to
copy the individual data blocks to, in the on-disk filesystem. Hence, the start
block is augmented to hold block addresses of the data-blocks in the
transaction. If the number of data blocks in the transaction exceeds the number
of addresses that can be stored in the start block, we add a dedicated address
block to the commit to hold the remaining block addresses (this is done only
on-demand). Further, scalefs uses a journal size of about 4 MB, which implies
that we will need at most one dedicated address block in any committed transaction.

Only the start block and the commit block have the transaction timestamp, which
helps us order dependent transactions during crash-recovery. Further, a
transaction is considered to be successfully committed only if the timestamps in
its start and commit blocks match. This also helps simplify journal-reset: when
the journal gets full, its transactions are applied to the on-disk filesystem to
make space for other transactions to commit to the journal. At this point, we
need to reset the journal so that the old transactions don't interfere with the
newly written transactions (which may include partial writes, if we crash
early). In earlier designs, we used to write a "zero" block at the head of the
journal before adding new transactions, but with this new design, journal-reset
doesn't involve any disk writes in the common case; we can simply reset the
write-offset to the head of the journal and immediately start committing the
next transaction.

This new design of the journal layout provides a 2x performance improvement with
parallel LFS smallfile benchmark on disk.