scalefs: Use a bitmap to keep track of disk id's to flush during ideflush()
Currently we maintain a vector of disk id's that every disk-write updates, and
later, just before calling ideflush(), this disknum vector is deduplicated so as
to invoke flush() exactly once on every disk that was previously written to.
However, this turns out to be extremely inefficient! The deduplication code
sorts the disk id's in order to discard the duplicates, and we suffer
quick-sort's worst case performance when the numbers are nearly sorted already,
which is unfortunately very common in our usecases. To give a specific example,
when using only 1 disk, we just have a single disk id (say 0), repeated
throughout the disknum vector, and quick-sort takes 1.9 seconds to "sort" 50,000
repeated zeros (when writing 200 MB of data to the disk), and about 17.8 seconds
to "sort" 204,800 repeated zeros (yielded by writing 400 MB of data to the
disk). In contrast, the actual disk writes take way less time!

Fix this train-wreck by maintaining a bitmap of the disk id's that need to be
flushed. This completely gets rid of the need to deduplicate the disk id's, and
is also much more memory-efficient.