ahci: Use FLUSH_CACHE_EXT to flush the cache when using 48-bit LBA addressing
The 48-bit address feature set in the ATA standard includes commands such as
READ DMA EXT and WRITE DMA EXT to read/write sectors on the hard disk.
The corresponding command to flush the volatile write-cache of the device
is FLUSH CACHE EXT (as opposed to FLUSH CACHE for 28-bit addressing mode).
Since we use 48-bit addressing in our driver, use the FLUSH CACHE EXT command
to flush out data to the non-volatile media.