ahci: Enable other interrupt types that signal various progress conditions
Currently we enable the Device-to-Host Register FIS interrupt (DHRE), which is
asserted when a data transfer command such as read/write has completed fully.
However, it turns out that when NCQ commands are issued, the DHRE interrupts are
not sufficient by themselves; the system hangs towards the end of booting
because some read/write doesn't complete, as it never gets the interrupt it is
waiting for.

Enabling some of the other interrupt types described in the AHCI spec fixes this
problem, so enable them in the driver. One reason why this could happen is that
some of these interrupts are "opportunistic" and can be missed if their timing
is unfortunate (this is explicitly documented in the AHCI spec). However, it is
still a mystery why this affects only NCQ commands and not the other ones.

Also, when enabling these additional interrupt types, we note that the
'Descriptor Processed' interrupt has a special requirement - we need to set the
'I' bit (bit 31 in the dbc) for every PRD entry, so as to cause these interrupts
upon processing the PRDs.