ahci: Add support to print out the NCQ Queued Error Log
SATA devices implement several log files which give detailed information about
various types of errors that occurred in the system. One among them is the NCQ
Queued Error Log, situated at log address 10h; this provides error information
for recently executed (failed) native queued commands.

This is quite valuable, because this is the only structure that tells us exactly
what the error was, as opposed to other sources such as the IS register
(Interrupt Status) or the SERR register (SATA Error) that only classify errors
into much more coarse-grained types. This debug feature was very useful when
developing NCQ support, so retain it in case we need it in the future.