disk: Provide vector APIs (namely scatter-gather lists) to read/write the disk
We already have low-level interfaces and mechanisms in place to fill up
read/write requests in terms of scatter-gather lists and submit it to the disk
using the AHCI driver. Expose this feature to higher layers (such as the
filesystem), and re-implement the scalar APIs in terms of the vector APIs.