boot, disk: Speed up booting by flashing the on-disk filesystem using writev()
Use scatter/gather I/O to flash the filesystem image to the disk faster, by
writing out the data in chunks of 64KB (instead of 4KB [block-size]).

Writing a 16 GB filesystem image to the disk used to take ~346 seconds before
this patch, and now takes only ~149 seconds (i.e., an improvement of 2.3x).