disk: Implement a simple block layer for ScaleFS, to achieve high performance I/O
Writing blocks individually to the disk (with a block size of 4KB) tends to be
rather inefficient, even when the blocks are contiguous. On the other hand,
using scatter-gather I/O to write contiguous blocks to the disk in bigger chunks
improves performance significantly.

Add a simple block layer below ScaleFS so as to accumulate contiguous
block-writes from the filesystem and use the disk driver's scatter-gather I/O
APIs to perform the disk writes efficiently. The block layer class, called
"block-queue", accepts write requests from higher layers and maintains a queue
of writes to contiguous block addresses. When they add upto a specified size
(64KB at the moment), it invokes the disk driver using the writev() API to write
the accumulated blocks to the disk. The block layer also transparently handles
I/O to multiple disks, using the preconfigured striping parameters to remap the
block offsets. An important feature of this design of the block layer is that it
does not impose any restrictions on its scope -- it can be instantiated any
number of times, and for "write-contexts" of arbitrary granularity. This means
that we can have a private block layer for every transaction, or a single global
block layer for the entire filesystem, or any combination thereof. This gives us
the flexibility to choose the option that works best, depending on the
contiguity of disk blocks in various scenarios.