scalefs: Use a private block layer per transaction to write out disk blocks
Add a private instance of the block layer to every transaction, so that the
transaction's blocks can be accumulated and written out to disk in bigger
chunks (via the block layer) to get higher disk throughput.

Having a common (shared) block layer that accumulates disk blocks from all
transactions can be potentially inefficient - blocks to be written to disk are
more likely to be contiguous within a given transaction (particularly when the
blocks are to be committed to an on-disk journal), whereas blocks across
different transactions may not be contiguous when they arrive interleaved at the
block layer, thus reducing the efficiency of vector writes. Hence, to avoid this
problem, we start out with a simple design where every transaction has a private
block layer that only queues its own blocks in order to perform scatter-gather
I/O to the disk.

However, note that not every transaction actually ends up writing to the disk,
for various reasons (group-transaction-commit being one of them). So take care
to instantiate the block layer only for transactions that actually intend to
write to disk, so as to save memory.