condvar, IPI: Implement faster cross-CPU task wakeups using asynchronous IPIs
Condition variables routinely need to wakeup blocked tasks whose wait-conditions
have been met. When such a task is to be woken up on the same CPU where the
condition variable's completion was signalled, the task gets woken up
immediately.  However, if the task is to be woken up on a different CPU (because
of the scheduling algorithm or CPU affinity settings), condition variables don't
achieve quick wakeups. They simply enqueue the task to the target CPU's
scheduling queue quietly and expect that CPU's scheduler to notice it on its
next scheduling run, which could be as far as 10ms away (the usual scheduling
time quantum). This causes noticeable delays in I/O bound jobs and leads to very
high idle times and poor performance.

Fix this by sending an IPI (Inter-Processor Interrupt) to the target CPU on
which the task is to be woken up, so that its scheduler will get to work
rightaway. Since the only reason we use an IPI here is to bring the target CPU
out of low-power idle state and begin its normal execution, the IPI handler
itself doesn't have to do anything! So just an asynchronous IPI with no payload
will suffice.

This improves the on-disk performance of LFS Smallfile on 2 CPUs by nearly 70%!
[ Before patch: 26.91 files/sec. After patch: 45.32 files/sec. ]