scheduler: Fix synchronization when dequeuing delayed work items
enq_dwork() holds the scheduler lock when enqueuing delayed work items to the
queue, but try_dwork() pops items off the queue without holding that lock. This
is unsafe, and also particularly dangerous because try_dwork() transfers CPU
control to the (potentially garbage) work item that it popped off the queue.

One manifestation of this bug occurs when there are a lot of calls to fork() and
wait() in the system. The wait() system call creates delayed work items to clean
things up, and the lack of synchronization in try_dwork() makes the instruction
pointer jump to god-knows-where! Debugging this was painful, because I had to
reconstruct the state of the stack by hand by looking at partial stack traces,
to arrive at the approximate location of the code (return address) before the
instruction pointer got corrupted beyond recognition.