oplog: Provide the option to use either a spinlock or a sleeplock for synchronization
The ScaleFS code needs to use a sleeplock to synchronize its oplog-managed
objects (because it does I/O while holding the lock), whereas the rmap code
needs to use a spinlock. However, the sync_lock_ in oplog is always a sleeplock
at the moment, which makes the rmap code deadlock-prone.

Fix this by allowing oplog's user to choose between a spinlock and a sleeplock,
based on its synchronization needs.