scalefs: Fix panic caused by op leftover on rename_stack due to benign fsync race
Flushing a cross-directory rename operation involves gathering both the parts of
the rename (i.e., rename-link and rename-unlink) from the two directories, and
then processing them together as an atomic unit. We achieve the former using a
2-step sequence in process_metadata_log(): first, we get a RET_RENAME_SUBOP
return code, followed by a RET_RENAME_PAIR (which we then process atomically).

However, when concurrent fsync()s are in flight on the same set of directories,
one of them can race ahead and process the 2-step rename sequence to completion,
while the other fsync() only manages to get as far as the first step
(RET_RENAME_SUBOP). In such cases, the latter fsync would expect a
RET_RENAME_PAIR next, but would instead get a RET_DONE, because there would be
nothing more to process. This ends up leaving unused metadata on the
rename_stack, causing the following assert to fail in process_metadata_log():

    assert(rename_stack.empty());

It doesn't actually matter which fsync() call flushes the operation, as long as
it is done correctly. So add logic under RET_DONE to fixup the rename_stack in
such cases, so that we don't leave stale metadata lying around.