tools/Makefile: Fix the flags used for compiling mkfs
We should be able to use macros to control filesystem configuration based on the
target build (an example is shown below):

#if defined(HW_qemu)
#define NINODES		X
#else
#define NINODES		Y
#endif

Unfortunately, this used to cause the kernel (filesystem) to pick up the value
X, while mkfs (userspace) used to pick up the value Y, because the "HW_" symbol
was not being defined when compiling mkfs. Fix it.