scalefs: Redesign per-core journaling for deadlock-freedom (Part 2/2)
This patch implements the code for applying the transactions committed in the
per-core journals, to their respective home locations on the filesystem.

Note that committed transactions are typically applied only when the per-core
journal gets full. Therefore, unlike the commit path, when trying to apply
transactions which have dependencies on different per-core queues, we can't
afford to simply wait (sleep) until all the dependent transactions have been
applied, because there might not be any processes driving the apply logic on
those other cores; so we risk ending up waiting forever.

So we take a more active approach instead: if CPU X depends on CPU Y to apply a
transaction, then CPU X grabs CPU Y's locks to forcefully apply the dependent
transactions to the filesystem. These dependencies can form a long (and rather
unpredictable) chain across different cores, so unless we are very careful in
acquiring the locks belonging to other cores, this can lead to a number of
deadlocks.

So the crucial design point in this redesigned code is to prohibit nested locks
in the apply path: locks belonging to different cores are never acquired in a
nested fashion, which automatically avoids lock-ordering related problems. So in
the above example, when CPU X wants to flush out CPU Y's apply-queue, it first
releases its own apply-lock, and *then* grab's CPU Y's apply-lock. Further, this
principle is honored even when following long chains of apply-dependencies
across different cores, which makes it deadlock-safe.