oplog: Make repeated calls to synchronize_upto_tsc() efficient
When processing entries from the oplog during an fsync(), we may end up calling
synchronize_upto_tsc() on the same oplog object multiple times, with the same
tsc value! So add some logic in oplog to avoid any extra work if we had already
synchronized upto that timestamp earlier.

Repeated calls to synchronize_upto_tsc() may seem redundant, but they are
actually used (or will be used) to acquire (and hold) the appropriate lock to
prevent concurrent updates to the object's synchronized value.