scalefs, oplog: Use spinlocks to manage scalefs objects within oplog
rmap and scalefs make use of the oplog in sv6. However, rmap uses spinlocks,
whereas scalefs uses sleeplocks in the core oplog code (sync-lock). Since both
of them share the same system-wide oplog cache, this discrepancy in locking
semantics can cause synchronization problems in certain scenarios, as explained
below.

oplog's get_logger() function will often need to evict an entry from the oplog
cache if there is a different object occupying the cache-slot that it wants to
use.  Performing this eviction involves acquiring the victim object's sync-lock.
Now consider the case where rmap is trying to evict a scalefs object from oplog.
Here, rmap uses a try-acquire loop to try and acquire the scalefs object's
sleeplock without actually sleeping. However, if the scalefs code path has
acquired that same lock and has gone to sleep (say, waiting for disk I/O or a
different sleeplock), then rmap will be stuck for a long time (with interrupts
disabled), trying to evict the scalefs object from the oplog cache. This causes
the watchdog to time out and complain that the system has locked up.

To fix this, we have already modified the scalefs code so as to not sleep while
holding its sync-lock. So now, we are finally ready to convert the sync-lock for
scalefs objects to also be spinlocks. This makes the oplog locking semantics to
be homogeneous for both rmap and scalefs, thus preventing any lockups during
cache-eviction.