scalefs: Send commit notifications _after_ post-processing the transactions
post_process_transaction() is used to release some of the resources held by the
transaction until it is safely committed to the on-disk journal. However, we
should try and release these resources as soon as we can, and certainly before
sending out the commit-notifications to the processes waiting for the commit to
finish, because technically the commit is truly complete only after the
post-processing step is also done.