scalefs: Make sure that commit- and apply-timestamps increase monotonically
On any given core, the values of the commit-timestamps and apply-timestamps of
transactions should always increase. In other words, notify_commit() (and
similarly notify_apply()) on a given core should never go back and forth between
the timestamps it announces as being completed. Add some assertions to verify
this property.