scalefs: Fix ordering of transactions to ensure monotonicity of timestamps
Once in a while, the statement that asserts that commit-timestamps on each core
should be monotonically increasing, fails. This is due to the race condition
depicted below. In short, the problem is the asymmetry between the per-core
journal's locks acquired by process_metadata_log() and sync_file().

Note that both the threads execute on the same core X. Also, commitq_insert_lock
is a sleeplock, whereas tx_commit_queue_lock is a spinlock.


        Thread 1 on CPU X                        Thread 2 on the same CPU X
--------------------------------------       ---------------------------------

      process_metadata_log():                          sync_file():

 - Acquire commitq_insert_lock

 - add_transaction_to_queue(trans)           - add_transaction_to_queue(trans)

     - trans->enq_tsc = get_tsc()
       (this gets a lower tsc value)
                                             - trans->enq_tsc = get_tsc()
                                               (this gets a higher tsc value)

                                             - Acquire tx_commit_queue_lock
                                             - tx_commit_queue.push_back(trans)
                                             - Release tx_commit_queue_lock

     - Acquire tx_commit_queue_lock
     - tx_commit_queue.push_back(trans)
     - Release tx_commit_queue_lock

 - Release commitq_insert_lock


As shown above, the transactions are NOT enqueued in the same order in which
they got their timestamps, which violates the monotonicity of timestamps of
transactions in the commit-queue.

Fix this by acquiring the per-core journal's commitq_insert_lock in the
sync_file() path as well.