scalefs: Hold commitq_remove_lock when checking the state of the commit-queue
In a recent patch, we fixed a nasty bug in the apply-code caused by observing
the state of the apply-queue without holding the applyq_remove_lock (for the
purpose of returning early from apply_all_transactions()). We have an almost
identical piece of code in the commit path as well, so "apply the same fix"
as a precaution.

However, as noted in the comments, there is no actual bug in the commit path
right now; we just want to make things symmetric, to rule out any potential for
future bugs arising from incorrect programmer assumptions due to this subtle
asymmetry.