scalefs: Fix inode leak due to lazy decrement of mnode refcount
In delete_mnum_inode_safe(), we have checks to ensure that we don't delete the
inodes of files that have open file-descriptors in user-space, even if they were
already unlinked from the filesystem hierarchy. This is based on checking the
true reference count of the mnode. However, when using mailbench (with
posix_spawn instead of fork) we observe a LOT of false positives -- cases where
the mnode refcounts are unusually high even for files that aren't supposed to
have any open file-descriptors; hence such inodes don't get deleted, thereby
causing rapid depletion of all the inodes in the system within a few benchmark
runs.  This is a transient problem, however; all the mnode refcounts *do* drop
to zero eventually, but just not quickly enough to delete the underlying inodes
in the fsync path.

Fix this by deleting the inode corresponding to the target mnode asynchronously:
i.e., when the mnode hits the onzero() function. That's the most reliable
indication that there are no open file-descriptors referring to that mnode, and
hence we can safely free the inode on the disk at that point. This fixes the
inode leak in mailbench.

Note that fsync() (after an unlink) can return right after the link-name is
deleted from the on-disk filesystem; it doesn't have to wait until the
underlying resources (inodes/blocks) are freed. So the asynchronous inode
deletion is free to take its time and run independently; it won't hold up
fsync().