sysfile, oplog: Fix logging of operations during early boot
We avoid logging filesystem operations in the oplog during early boot, as they
involve things like creating device special files etc., which we don't want to
get entangled in the sync/fsync path. However, we missed a few places in
create() where this should be checked, so fix it. Without this, it causes an
imbalance between start_tsc and end_tsc in oplog's data-structures, which leads
to mysterious bugs.