mailbench: Add support for per-cpu spool directories and multiple user mailboxes
This adds support to create per-cpu spool directories and spawn an instance of
mail-qman to manage each spool directory. This also supports multiple user
mailboxes (currently limited to the same number as the number of spool
directories).

For example,

$ ./mailbench -c -u 4 basedir 4

starts mailbench with 4 spool directories and 4 user mailboxes, and creates a an
independent set of mail-* processes to manage each pair of spool- and user-
directories.