scalefs: Remove the mechanism to mark/unmark unreachable inodes
We have a mechanism for marking/unmarking a newly created file or directory as
potentially unreachable, depending on whether its parent directory was fsynced
after creating the new file or directory in it. This helps speed up the process
of reclaiming orphaned inodes and their disk blocks upon reboot, by having a
ready list of possible suspects.

There is a bug in this code that causes occasional crashes, but instead of
fixing it, we can actually remove it altogether! We already read all the inodes
off the disk when booting, so as to create an in-memory bitmap of the inodes
that are free; so, using this, we can simply look at the nlink field of each
inode during boot, and reclaim it if nlink is 0. We don't really need another
mechanism to keep a list of orphaned inodes.

Our current filesystem configuration has 1 million inodes, and the time it takes
to read them all during boot is almost negligible (0.7s on an SSD). So
piggybacking inode-reclaim on this mechanism seems acceptable. As an added
advantage, this avoids a few scattered disk-writes when fsyncing each newly
created file or directory.