mnode, onzero: Don't flush the transaction queue when deleting the inode
When an mnode's onzero() function is reached and it is marked for deletion, we
construct a transaction that deletes its on-disk inode. However, we don't need
to actually flush the transaction to the on-disk journal and commit it rightaway;
it can be done lazily by a subsequent sync/fsync, possibly reaping the benefits
of group-commit as well. This avoids potentially performing a separate commit
for every inode deletion!