scalefs: Delete on-disk inodes primarily through the mnode::onzero() path
Currently we try to immediately delete the on-disk inode corresponding to an
mnode if its reference count has dropped below a certain value. However, we can
also simply postpone the delete until the mnode's onzero() function is called,
which is already being used in some situations. So make that the primary way of
deleting on-disk inodes.

Besides, refcache's get_consistent() call has never really worked correctly
anyway, and furthermore, it is expensive too, because it has to accumulate the
actual count by reading all the per-core refcounts, thus causing cache-line
conflicts. This patch gets rid of the call to get_consistent() from ScaleFS.