scalefs, block-allocator: Fix performance issue related to I/O coalescing
When running lfs-largefile (just the sequential 100 MB write + fsync) on an SSD,
Linux ext4 takes 0.5 seconds, and ScaleFS takes 0.54 seconds.

However, the performance of ScaleFS was not reproducible after the first run --
i.e., the second run of lfs-largefile after booting, and every subsequent run,
took 0.91 seconds, which is a 70% slowdown!

The issue turned out to be simple, yet rather amusing: in the first run, the
blocks for the 100 MB file are allocated sequentially, in increasing order, from
the corresponding CPU's block-allocator. But when the benchmark completes, it
returns the blocks in the same order (first to last) one by one, which is added
to the head of the block freelist. So on a subsequent run, the blocks allocated
will still be sequential, but in the reverse order (block n, block n-1 etc)!

The reason this matters for performance is that we have a block layer under the
filesystem that tries to coalesce contiguous I/O upto 64KB chunks and hand them
over to the disk driver for scatter-gather I/O. This provides significant
performance boosts as opposed to doing I/O in 4KB chunks. However, when the
blocks written are contiguous but in the reverse order, the block layer does not
coalesce them because they won't be amenable to scatter-gather I/O (unless they
are explicitly sorted; but we want to avoid sorting potentially thousands of
blocks at that point).

Thus, the second and subsequent runs of lfs-largefile would fail to reap the
benefits of I/O coalescing, causing the slowdown. Fix this by adding freed
blocks to the end of the freelist, so that we can allocate them in increasing
order in subsequent runs as well.