scalefs, mnode: Fix deadlock caused by skipping transaction flush from mnode::onzero()
As an optimization, commit f7fd9476a decided not to flush the transaction queue
after enqueuing a transaction to delete inodes from mnode::onzero(). While this
is correct in principle, the journaling code has slightly different
expectations: it expects that any new transaction that is enqueued on a per-core
transaction queue is driven all the way up to the point of being committed to
that core's on-disk journal. That way, when there are cross-queue dependencies
between transactions, one simply has to wait for the appropriate transaction to
get committed, before proceeding to commit the dependent transaction; in
particular, we don't have to poke the other transaction queue or force-flush its
transactions to disk.

Commit f7fd9476a broke this assumption by enqueuing a transaction and just
leaving it in the queue, instead of seeing it through all the way upto commit.
As a result, it causes the code to hang whenever a transaction from a different
per-core queue ends up waiting for this (abandoned) transaction to commit.

Fix it by having a different scheme to implement the optimization: let
mnode::onzero() note down the inode number to be deleted in a per-core
data-structure, so that a subsequent sync or fsync on that core can delete those
inodes and also drive them up to the commit point.