scalefs: Reduce the number of disk-flushes per commit by using checksums
Currently we perform 2 disk-flushes for every transaction commit:

 - Write the journal header start block and all the datablocks of the
   transaction.
 - Issue a disk flush.
 - Write the journal header commit block.
 - Issue another disk flush.

This was necessary to ensure the integrity of each transaction commit, when only
using the ordering of disk writes to guarantee correctness.

However, disk-flushes are expensive, so this patch uses checksums to reduce the
number of disk-flushes per commit to 1. So, the new scheme for commit is:

 - Write the journal header start block and all the datablocks of the
   transaction.
 - Write the journal header commit block, which now includes the checksum
   (CRC16) of all the datablocks in the transaction.
 - Issue a disk flush.

The crash-recovery code can then use the checksum to verify if the transaction
was committed successfully, and if so, apply it to the filesystem.

This causes mailbench single core performance on 1 SSD to increase from 18
messages/sec to 23 messages/sec.