scalefs: Fix deadlock in sync() caused by queueing to different per-core journals
In the sync() path, we flush each per-core transaction queue completely one
after another. So it is important that we don't add dependent transactions
(those that were generated by us in the sync path) to different per-core
transaction queues, because then we wouldn't be able to resolve
wait_for_commit(). So avoid this deadlock by committing all transactions in the
sync path via our core's journal.