bin/lfs-largefile: Move to attic
Going forward, we want to make largefile run multi-threaded, among other
enhancements. But this code is a dinosaur! It uses a very old style of C and is
quite inflexible. So dump it in the attic. We will use the lfs-smallfile code as
a base to rewrite the largefile benchmark. The smallfile code already has
support for running parallel threads/processes etc., so we can reuse all that
code.