memide: Make arrangements to allocate the ramdisk in discontiguous chunks
At the moment, we allocate the memory for the ramdisk as one giant contiguous
chunk of memory (nearly 16GB) at early boot, to accommodate the entire
filesystem. However, this turns out to be a bad idea for performance because all
of this memory gets allocated from a single numa node.

So going forward we would like to allocate memory for different parts of the
ramdisk by spreading the allocations across the available numa nodes, so as to
prevent a single numa node from becoming the hotspot for all disk writes via
memide. To facilitate this, make arrangements to allocate the ramdisk at the
granularity of a block (BSIZE), and use an array of pointers to redirect the
reads/writes appropriately.